requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pydantic
python-jose[cryptography]
